// src/types/collections.d.ts
export type OrderStats = {
  weightedAverage: string
  max: string
  min: string
  stddev: string
  median: string
  volume: string
  orderCount: string
  percentile: string
}

export type MarketCache = {
  typeId: number
  structureId: string
  buy: OrderStats
  sell: OrderStats
}

export type MarketCacheDocument = WithID<MarketCache>